Rename `workflow` configurations module to `basic`
