+++ src/aiidalab_qe/app/configuration/basic/__init__.py
@@ -1,5 +1,5 @@
+from .basic import BasicConfigurationSettingsPanel
 from .model import BasicConfigurationSettingsModel
-from .workflow import BasicConfigurationSettingsPanel
 
 __all__ = [
     "BasicConfigurationSettingsModel",
